Fix socket issue
Omit bind/listen steps (performed by systemd activation)
and rebind the socket

+++ tendrl/node_agent/message/handler.py
@@ -17,6 +17,7 @@ from tendrl.commons.message import Message
 from tendrl.commons.logger import Logger
 
 
+LISTEN_FDS = int(os.environ.get("LISTEN_FDS", 0))
 MESSAGE_SOCK_PATH = "/var/run/tendrl/message.sock"
 
 
@@ -46,7 +47,7 @@ class MessageHandler(gevent.greenlet.Greenlet):
             exc_type, exc_value, exc_tb = sys.exc_info()
             traceback.print_exception(
                 exc_type, exc_value, exc_tb, file=sys.stderr)
-            
+
     def _read(self, sock, size):
         data = ''
         while len(data) < size:
@@ -55,12 +56,12 @@ class MessageHandler(gevent.greenlet.Greenlet):
             if dataTmp == '':
                 raise RuntimeError("Message socket connection broken")
         return data
-    
+
     def _msgLength(self, sock):
         d = self._read(sock, 4)
         s = struct.unpack('=I', d)
         return s[0]
-    
+
     def _run(self):
         try:
             self.server.serve_forever()
@@ -75,27 +76,33 @@ class MessageHandler(gevent.greenlet.Greenlet):
     def bind_unix_listener(self):
         # http://0pointer.de/blog/projects/systemd.html (search "file
         # descriptor 3")
-        try:
-            socket_fd = 3
-            self.sock = socket.fromfd(socket_fd, socket.AF_UNIX,
-                                      socket.SOCK_STREAM)
-            self.sock.setblocking(0)
-            self.sock.listen(50)
-            return self.sock
-        except (TypeError, BlockingIOError, socket_error, ValueError):
-            exc_type, exc_value, exc_tb = sys.exc_info()
-            traceback.print_exception(exc_type, exc_value, exc_tb,
-                                      file=sys.stderr)
-            pass
-        try:
-            self.sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-            if os.path.exists(MESSAGE_SOCK_PATH):
-                os.remove(MESSAGE_SOCK_PATH)
-            self.sock.setblocking(0)
-            self.sock.bind(MESSAGE_SOCK_PATH)
-            self.sock.listen(50)
-            return self.sock
-        except:
-            exc_type, exc_value, exc_tb = sys.exc_info()
-            traceback.print_exception(exc_type, exc_value, exc_tb,
-                                  file=sys.stderr)        
+        if LISTEN_FDS == 0:
+            # Create new socket
+            try:
+                socket_fd = 3
+                self.sock = socket.fromfd(socket_fd, socket.AF_UNIX,
+                                          socket.SOCK_STREAM)
+                self.sock.setblocking(0)
+                self.sock.listen(50)
+                return self.sock
+            except (TypeError, BlockingIOError, socket_error, ValueError):
+                exc_type, exc_value, exc_tb = sys.exc_info()
+                traceback.print_exception(exc_type, exc_value, exc_tb,
+                                          file=sys.stderr)
+                pass
+            try:
+                self.sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+                if os.path.exists(MESSAGE_SOCK_PATH):
+                    os.remove(MESSAGE_SOCK_PATH)
+                self.sock.setblocking(0)
+                self.sock.bind(MESSAGE_SOCK_PATH)
+                self.sock.listen(50)
+                return self.sock
+            except:
+                exc_type, exc_value, exc_tb = sys.exc_info()
+                traceback.print_exception(exc_type, exc_value, exc_tb,
+                                          file=sys.stderr)
+        else:
+            # Rebind Socket
+            # Omit bind/listen steps (performed by systemd activation
+            self.socket = socket.fromfd(3, socket.AF_UNIX, socket.SOCK_STREAM)
